env환경변수 셋팅
- joi를 사용해서 configuration 셋팅 후 configservice로 환경변수를 사용할 수 있음
