+++ server/.env
@@ -1,0 +1,7 @@
+DB_TYPE=postgres
+DB_PORT=5432
+DB_HOST=localhost
+DB_USERNAME=jeongdaun
+DB_PASSWORD=ekdns001
+DB_SCHEMA=users
+
