+++ server/src/api/v1/users/users.entity.ts
@@ -30,7 +30,4 @@ export class Users {
 
   @Column()
   phone_number!: string;
-
-  @Column()
-  mbti?: string;
 }
