+++ server/src/config/config.validator.ts
@@ -1,0 +1,14 @@
+import { ConfigModuleOptions } from '@nestjs/config';
+import * as Joi from 'joi';
+
+export const CONFIG_VALIDATOR: ConfigModuleOptions = {
+  validationSchema: Joi.object({
+    DB_PORT: Joi.number().required(),
+    DB_HOST: Joi.string().required(),
+    DB_USERNAME: Joi.string().required(),
+    DB_PASSWORD: Joi.string().required(),
+    DB_SCHEMA: Joi.string().required(),
+  }),
+  isGlobal: true,
+  cache: true,
+};
